+++ plugins/calibre-library/__init__.py
@@ -12,7 +12,7 @@
 from calibre.customize import InterfaceActionBase
 
 PLUGIN_NAME = 'epubveri library'
-PLUGIN_VERSION_TUPLE = (0, 4, 0)
+PLUGIN_VERSION_TUPLE = (0, 5, 0)
 PLUGIN_VERSION = '.'.join(str(part) for part in PLUGIN_VERSION_TUPLE)